rcytoscapejs {rcytoscapejs}	R Documentation

Create CytoscapeJS widget

Description

Pass network information to CytoscapeJS widget

Usage


rcytoscapejs(nodeEntries, edgeEntries, layout = "cose", width = NULL,
  height = NULL, showPanzoom = TRUE, highlightConnectedNodes = TRUE,
  boxSelectionEnabled = TRUE)

Arguments

nodeEntries
nodes output from createCytoscapeJsNetwork function

edgeEntries
edges output from createCytoscapeJsNetwork function

layout
the layout algorithm to use for layout. If 'preset', then columns 'x' and 'y' from nodeEntries will be used for position information.

width
width of widget (Default: 100 percent for Shiny apps)

height
height of widget (Default: 400px for Shiny apps)

showPanzoom
whether to show the CytoscapeJS panzoom menu in the upper left corner (Default: TRUE)

highlightConnectedNodes
whether to highlight connected nodes to a clicked node (Default: TRUE)

boxSelectionEnabled
whether to allow users to box select multiple nodes (Default: TRUE)

See Also

createCytoscapeJsNetwork

Examples


id <- c("Jerry", "Elaine", "Kramer", "George")
name <- id
nodeData <- data.frame(id, name, stringsAsFactors=FALSE)

source <- c("Jerry", "Jerry", "Jerry", "Elaine", "Elaine", "Kramer", "Kramer", "Kramer", "George")
target <- c("Elaine", "Kramer", "George", "Jerry", "Kramer", "Jerry", "Elaine", "George", "Jerry")
edgeData <- data.frame(source, target, stringsAsFactors=FALSE)

network <- createCytoscapeJsNetwork(nodeData, edgeData)
rcytoscapejs(network$nodes, network$edges)


[Package rcytoscapejs version 0.0.7 Index]