renderRcytoscapejs {rcytoscapejs}	R Documentation

Widget render function for use in Shiny

Description

Widget render function for use in Shiny

Usage


renderRcytoscapejs(expr, env = parent.frame(), quoted = FALSE)

Arguments

expr
An expression that returns an R object that can be used as an argument to cat.

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

[Package rcytoscapejs version 0.0.7 Index]